requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
dj-database-url==1.0.0
Django==4.0
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-js-asset==2.0.0
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.4
multidict==6.0.2
Pillow==9.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
yarl==1.8.1
